organize files and folders and add media query for the map
