+++ src/components/Location/Location.css
@@ -1,0 +1,33 @@
+.Location {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    font-family: Futura;
+}
+.Location h4 {
+    color: #CB8B4E;
+    letter-spacing: 2px;
+    font-weight: 500;
+    margin-bottom: 10px;
+}
+.Location h2 {
+    margin-top: 10px;
+    margin-bottom: 40px;
+    color: #333333;
+    font-weight: 500;
+    font-size: 56px;
+}
+#map {
+    height: 180px;
+}
+.leaflet-container {
+    width: 70vw;
+    height: 70vh;
+}
+
+@media screen and (max-width: 480px) {
+    .leaflet-container {
+        width: 100vw;
+        height: 50vh;
+    }
+}
